requirements.txt
dspy==2.5.29
python-dotenv==1.0.1
litellm==1.52.6
